Move range parsing into separate module
Make it more generic, for future range parsing.
